arreglo de reporte

+++ database/migrations/2022_05_12_052839_create_reporte_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
         Schema::create('reporte', function (Blueprint $table) {
             $table->engine="InnoDB";
             $table->bigIncrements('id_reporte');
-            $table->int('id_usuario')->references('id')->on('users')->onDelete("cascade");
+            $table->bigInteger('id_usuario')->references('id')->on('users')->onDelete("cascade");
             $table->date('fecha_reporte');
             $table->timestamps();
         });
